Fixed incorrect cases in VertexBuffer layout
Fixed incorrect cases in VertexBuffer layout

+++ Sandbox/assets/shaders/Unlit2D.glsl
@@ -6,7 +6,7 @@
 layout(location = 0) in vec3 a_Position;
 layout(location = 1) in vec4 a_Color;
 layout(location = 2) in vec2 a_TexCoord;
-layout(location = 3) in float a_TexIndex;
+layout(location = 3) in int a_TexIndex;
 layout(location = 4) in vec4 a_Tiling;
 
 uniform mat4 u_ViewProjection;
@@ -21,7 +21,7 @@ void main()
 {
 	v_Color = a_Color;
 	v_TexCoord = a_TexCoord;
-	v_TexIndex = int(a_TexIndex);
+	v_TexIndex = a_TexIndex;
 	v_Tiling = a_Tiling;
 	gl_Position = u_ViewProjection * vec4(a_Position, 1.0);
 }
